+++ Changelog.md
@@ -1,7 +1,12 @@
+# 0.0.4
+- Add `bitmaskWithRejection32'` and `bitmaskWithRejection64'`
+  which generate numbers in closed range `[0, n]`.
+  Unticked variants generate in closed-open range `[0, n)`.
+
 # 0.0.3
 
 - Add `System.Random.SplitMix32` module
-- Add `bitmaskWithRejection32` and `bitmaskWithRejection64` module
+- Add `bitmaskWithRejection32` and `bitmaskWithRejection64` functions
 - Add `nextWord32`, `nextTwoWord32` and `nextFloat`
 - Add `random` flag, dropping dependency on `random`
   (breaks things, e.g. `QuickCheck`, when disabled).